Refactored Composables and removed unused resources.
